No longer fail when we do not get a timestamp from the HW
On a receive when using TCP there is not always a timestamp since
there is no correlation between receiving something and a packet.
In this case reuse the old timestamp we previously already took.

Also, ignore first packets that do not have a timestamp since
this will result in timestamp 0.

Signed-off-by: Roni Häcki <roni.haecki@inf.ethz.ch>